requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
